requirements.txt
Langchain
langchain-community
Langchain-core
langchain-text-splitters
langchain-openai
pypdf
faiss-cpu
Langchainhub
langsmith
Langgraph
langchain-pinecone
requests
black
openai
google-search-results
python-dotenv
pinecone
beautifulsoup4